moved Game utils to gameUtils.js file. Issue #6
